+++ src/main/java/jp/igapyon/cityinfojp/dyn2static/Dyn2StaticPrefProcessor.java
@@ -24,9 +24,9 @@ import org.thymeleaf.context.Context;
 import org.thymeleaf.context.IContext;
 import org.thymeleaf.spring5.SpringTemplateEngine;
 
-import jp.igapyon.cityinfojp.dyn.thymvarmap.ThymVarMapPrefBuilder;
 import jp.igapyon.cityinfojp.input.entry.PrefEntry;
 import jp.igapyon.cityinfojp.input.entry.PrefEntryUtil;
+import jp.igapyon.cityinfojp.thymvarmap.ThymVarMapPrefBuilder;
 
 public class Dyn2StaticPrefProcessor {
