+++ src/main/java/jp/igapyon/cityinfojp/thymvarmap/ThymVarMapPrefBuilder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package jp.igapyon.cityinfojp.dyn.thymvarmap;
+package jp.igapyon.cityinfojp.thymvarmap;
 
 import java.io.BufferedReader;
 import java.io.IOException;
