+++ src/main/resources/static/about.html
@@ -181,7 +181,7 @@
     </main>
 
     <footer class="container">
-        <p>&copy; cityinfojp 2020 : 2020-04-18 22:52</p>
+        <p>&copy; cityinfojp 2020 : 2020-04-18 23:24</p>
     </footer>
 
     <!-- Bootstrap JavaScript -->
