+++ src/main/resources/static/index.html
@@ -1793,7 +1793,7 @@
     </main>
 
     <footer class="container">
-        <p>&copy; cityinfojp 2020 : 2020-04-18 22:52</p>
+        <p>&copy; cityinfojp 2020 : 2020-04-18 23:24</p>
     </footer>
