+++ src/main/resources/static/sitemap.xml
@@ -2,7 +2,7 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
 <url>
 <loc>https://cityinfojp.herokuapp.com/</loc>
-<lastmod>2020-04-18T22:52:23+09:00</lastmod>
+<lastmod>2020-04-18T23:24:00+09:00</lastmod>
 <changefreq>daily</changefreq>
 <priority>0.8</priority>
 </url>
